+++ lib/features/bottomNavigationBar.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+// ignore: camel_case_types
 class bottomNavigationBar extends StatelessWidget {
   const bottomNavigationBar({Key? key}) : super(key: key);
