+++ pubspec.yaml
@@ -28,6 +28,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  carousel_slider: ^4.0.0
+  http: ^0.13.3
 
 dev_dependencies:
   flutter_test:
